Separate Redis cache and driver implementation (#194)
